+++ src/services/index.ts
@@ -2,6 +2,5 @@ export * from './toast';
 export * from './authCredentials';
 export * from './storage';
 export * from './searchHistory';
-export * from './cameraRoll';
 export * from './permission';
 export * from './multimedia';
