+++ src/services/multimedia/index.ts
@@ -1,2 +1,4 @@
 export * from './multimediaService';
 export * from './multimediaType';
+
+export * from './useMultimediaGetPhotos';
