+++ src/services/multimedia/multimediaType.ts
@@ -1,3 +1,9 @@
+export type PhotoListPaginated = {
+  photoList: string[];
+  cursor?: string;
+  hasNextPage: boolean;
+};
+
 export type ImageForUpload = {
   uri: string;
   name: string;
